refactor: moved HTTPError into errors.py
Makes more sense to move all custom error classes
into errors.py
