+++ tests/test_request.py
@@ -3,7 +3,7 @@ import pytest
 from requests import RequestException
 from unittest.mock import Mock
 
-from dimo.request import Request, HTTPError
+from dimo.errors import Request, HTTPError
 
 
 class DummyResponse:
